[dev] Service 계층 인터페이스와 구현체로 분리 (#13)
